Fixes #12754 - Fix deadlock in libcec initialization.
Configuring the TV at startup is done from the main thread. However, when the
TV state changes as a result of this, libCEC sends a message to the callback
handler on its own thread, which tries to get the UI lock to reset the
screensaver state. This results in a deadlock because the main thread already
has the UI lock during startup time.

Instead, run the initial setup activity at the start of the CEC handler thread.

This was missed in the patch applied for #11986.

Signed-off-by: Stuart Auchterlonie <stuarta@mythtv.org>
(cherry picked from commit 22ff48d7bafc622246141094b727dab47662bc3f)