musicscanner: Compare music_directory.path as binary
SQL string comparisons are case insensitive by default (who knew!), meaning
that a music rescan will not pickup a renamed directory if it differs only in
the case.

Use BINARY rather than arranging for a case-sensitive comparison because the
latter still considers 'a' and 'á' equal but for paths we need literal binary
sameness.

Apply the same logic to music_songs.filename, although I've not tripped over a
related issue there.

Fixes #12949.

Signed-off-by: Paul Harrison <pharrison@mythtv.org>