Allow multiple element names with different dependancies.
MythUI only allows one element with a given "name", so allow a themer to
specify names with an extension to make them unique:  Anthing after (and
including) a '+' will be ignored when matching the name.  For example:

<textarea name="description+sub" depends="subtitle">

and

<textarea name="description+notsub" depends="!subtitle">

results in the appropriate definition for "description" to be used depending
on if the "subtitle" element is available.

I thought about making the "+dependancies" naming automatic, but if the
themer is using inheritance, they may find the result unexpected.

This currently only works with textareas.  It would be quite a bit more work
to make it work with other types, and I have not been able to think of any
benifit.

(cherry picked from commit d82629a4dfe6d8b2baa86aa7e71b34e76638f536)