Fix compiler 'enumeral and non-enumeral type in conditional expression' warning.
The complaint is because the number zero is used in conjunction with
enum values.  Create two new enums, kDupsUnset and kDupCheckUnset, that
have the value of zero and use these instead.  Rewrite both the
RecordingInfo::DeleteHistory and MythMainWindowPrivate::TranslateKeyNum
functions to use these new enums so that the compiler sees only enum
values and doesn't complain.