Fix unused parameter warning for explicit dummy variable.
This function explicitly names its argument as 'dummy'. It would be
easy to remove the argument from the function and its sole caller as
they are all contained in the same file, but its a back pointer to the
VideoList object that is calling this VideoListImp function. Leave it
and just comment out the name.