Use XMLTV dd_progid data to create seriesid
The current seriesid processing for XMLTV grabbers using a hash for the
title is non-optimal in the (Schedules Direct) case where the seriesid can
be more accurately represented (via a supplied dd_progid). The internal
datadirect grabber creates such a seriesid from its feeds.

From some research and discussions with Schedules Direct regarding the
gracenote upstream (raw) data, there is a true seriesid available, but
passing that through to MythTV, and then adjusting existing data has some
challenges (some upstream). And while there are some potential approaches,
rather than wait for the perfect, we can do something now which is
equivalent to the internal dd grabber.

This patch creates a better seriesid from a XMLTV source which provides a
dd_progid with a seriesid which is compatible with the (internal)
DataDirect? grabber (further moving forward to allowing the eventual removal
of the internal DD grabber).

The following patch (based on some initial work by Mr. Poet) creates an
appropriate seriesid.

Note:

The uniqueid value was added to the code in commit
347ea0319330cb06cf3e418e12e79602c3235bc0 (over a decade ago) however it was
never set, and always empty, resulting in some dead code. It appears that
the uniqueid may have been intended to serve the same/equivalent purposes of
programid (at some point). This patch removes the (dead) uniqueid along the
way as part of the larger set of fixes.

As this patch changes the seriesid for new grabbed programs, those that are
using the "this series" matching based on the previous seriesid may run into
some issues. Likely the easiest solution to recommend is to recreate those
rules.

Signed-off-by: John Poet <jpoet@mythtv.org>
(cherry picked from commit 4e5cf2b4ef875ee578b121be9867ba05da5fab17)