VBOX: Add multi-record support to the vbox recorder
This allows the vbox recorder to record more than one channel using a single
tuner so long as they are on the same multiplex.

You need to uncheck 'Limit single channel streaming per UPnP device' setting
on the Vbox for this to work and firmware >= 2.53 is recommended.

Also Myth needs to know which channels share a multiplex so a retune of the
Vbox recorder in Myth is necessary in order to populate this information into
the database.

This patch also contains work arounds for #12856 and #12773.

Signed-off-by: Paul Harrison <pharrison@mythtv.org>