Adjust GuideGrid widget for the parent offsets.
This patches the UI GuideGrid widget to translate drawing positions
from coordinates that are local to the parent object into global
coordinates. The parents xoffset/yoffset values were already being
passed into this widget. These offsets just needed to be passed down
to the actual drawing functions, and those functions needed to perform
the additional translation. The lack of these translations has not
been noticed because all GuideGrids are currently embedded in top
level windows, so that the parent offsets are always 0/0.