Fix 'unused variable' warning in VideoOutputOMX::UseBuffersCB.
The assert macro compiles differently when debugging is enabled in the
'configure' command.  This means that the variable 'def' may or may
not be used depending on how the build is configured. This patch uses
the debug conditional compilation flag (the same one that assert uses)
so that the variable only exists when debugging. This will prevent the
compiler form complaining that it is an unused variable.