Support for IPV6 link-local addresses.
Fix support for link-local in startup window.
Create a new PortChecker class that checks if a port is open and
also resolves IPV6 link-local scope.
Change backend connection to use the PortChecker class.
Fix invalid xml in services when a link-local address with scope is sent.
No longer store the scope with the link-local ip address in config.xml or database
settings. If they are present they will be stripped and replaced with
correct value.

After all of this, IPV6 link-local addresses can be used for the database
connection in cofig.xml and for the backend ip address and master backend address.
Auto discovery of backend works for database with link-local address
but not if the backend also uses link-local listen address.

Still not working - services API with link-local address, because
http URLs with % sign in the IP address are not supported.