ChannelBase::GetChanID()   : Prefer a visible channel DVBChannel::GetChanID      : Prefer a visible channel ChannelUtil::GetChannelData: Prefer a visible channel
The Guide shows visible channels, and returns the associated chanid.  When
TVRec::StartRecording checks if the input is busy, though, it just pulls the
first chanid available for the channum.  This results in conflicts and an
existing recording can end up being killed by a new one, unnecessarily.

For example, with Directv, mythconverg.channel has:

+--------+---------+----------+---------+
| chanid | channum | callsign | visible |
+--------+---------+----------+---------+
|   2204 | 204     | HLN      |       0 |
|  11711 | 204     | HLNHD    |       1 |
+--------+---------+----------+---------+

The recording rule is created with 11711, but TVRec::IsBusy calls
channel->GetChanID() which was just returning the first chanid it found
(2204).