Overhaul sorting of strings.
Create the new files mythsorthelper.{h,cpp} to centralized all the
code related to sorting, ignoring case, handling prefixes, etc.  This
code handles the case where the translation says there are no prefixes
for a language.  It also handles a list of prefixes that should never
be sorted.  The main object here is essentially a singleton, but its
reference counted so that any time the preferences are changed it can
be tossed and a new 'singleton' created.

Rip out the Videometadata SortData and SortKey structures and replace
them with a simple "sortBefore" function.

Add sorting to mythmusic, mythnews, etc.

Remove unnecessary doxygen function names.

Fixes #12298.