Validate formats passed to get_format() functions.
According to a comment in external/FFmpeg/libavcodec/avcodec.h, the
last argument to get_format() is a list of valid formats, and the
behavior is undefined if the function doesn't return a value from the
supplied list. This patch makes the variable name more informative,
and runs the format list to ensure that the desired value is
present.