Fix for LCD device code hanging
When playing recordings if the audio changes (e.g. number of
channels change), it sends a message to the lcd server
from the wrong thread.
After this no LCD messages get through, but queue indefinitely.
The fix uses a slot to ensure messages are send from the
correct thread.

Refs #12566
Fixes #12567