Cleanup QObject related warnings for RecStatus.
The RecStatus object is based on QObject and therefore the object
itself cannot be used with the Qt MetaType system.  However, the
pointer to any class derived from QObject is automatically registered
with the MetaType system.  QObject derived classes are also not
allowed to have a copy constructor, so remove that as well.  This
allows all of the declaration/registration code to be removed from the
file.  This also adds in the Q_DISABLE_COPY macro to the class.

Instances of QObject (and its subclasses) should be considered unique
items that cannot be copied.  Knowing that each object has only one
parent and an arbitrary number of children makes it easier to
understand this. Even if you "clone" the object by setting all the
member variables to the same values, you can't set the parent and
child variables to the same values without screwing up the one-to-many
relationship of parent to children.

See:
http://doc.qt.io/qt-5/qobject.html#Q_DISABLE_COPY
https://www.ics.com/designpatterns/book/qobject.html