New feature for mythfilldatabase to optionally not use allatonce.
The option added to mythfilldatabase is '--no-allatonce' and is (sort of)
the equivalent to a no-dd-grab-all option (if it existed), --dd-grab-all
not being the default due to some systems being under-provisioned.

While for many grabbers that specify a preference for grabbing
"allatonce" using it has advantages, for grabbers that return a
large amount of data (example: 500+ channels for 21+ days with high
details) mythfilldatabase due to its internal structure of parsing
the entire xmltv file before processing can end up consuming vast
amounts of virtual memory (6GB+ in one test case), which some backends
may not be able to support well.

This patch adds in a --no-allatonce option to mythfilldatabase which
will cause the fill processing to fall back to a day at a time grab/fill,
trading off invoking the grabber multiple times (with increased CPU
procesing workload) for a reduced memory requirement.

Fixes #12758

Signed-off-by: Peter Bennett <pbennett@mythtv.org>