Use the 'block' argument to SendReceiveStringList().
Modify the SendReceiveStringList argument to use the 'block'
argument. If this argument is set to false and the backend connection
fails, then the function will not attempt to reconnect to the
backend. If block is true (the default) then the function will attempt
re-connection.

All current callers of this function (except one) leave the block
argument at its default value of true. The one caller that sets block
to false is never referenced anywhere.