Create a framework for testing data contracts.
This is a start on a test framework for the data contracts.  It
currently builds five different types of objects, converts them to a
QVariant and back, and then compares the results to the original.

This should be extended to support all the contracts.

Are there other tests that should be added?