TV: Fix ESCAPE being ignored
The cause was that lcdVolumeTimerId was never set to 0 after killing it.
Since timer IDs are reused by Qt the timer started when ESCAPE was
pressed to exit playback could be treated as lcdVolumeTimerId and ending
up never being processed.  This only happened if the LCD support was
enabled and the volume was changed.

Fixes #12566

(cherry picked from commit bdbc6363a311f08cdb61eab81d62243b56b7bd12)