Remove public copy constructors from MythSystemLegacy.
The MythSystemLegacy object is based on QObject and therefore
shouldn't have a public copy constructor. Update the code to
properly initialize the parent QObject, and to remove the
unused copy constructor.

General: Instances of QObject (and its subclasses) should be
considered unique items that cannot be copied.  Knowing that each
object has only one parent and an arbitrary number of children makes
it easier to understand this. Even if you "clone" the object by
setting all the member variables to the same values, you can't set the
parent and child variables to the same values without screwing up the
one-to-many relationship of parent to children.

See:
http://doc.qt.io/qt-5/qobject.html#Q_DISABLE_COPY
https://www.ics.com/designpatterns/book/qobject.html

Refs #13069