Cleanup QObject related warnings for the data/service contracts.
All the data/service contract objects are based on QObjects and
therefore the object itself cannot be used with the Qt MetaType
system.  However, the pointer to any class derived from QObject is
automatically registered with the MetaType system. This allows all of
the declaration/registration code to be removed from the files.
Removing all the registrations creates a significant cleanup, as all
of the InitializeCustomTypes functions go away.  Furthermore, now that
there are no InitializeCustomTypes functions, there is significant
cleanup to all of the service object constructors. This also adds in
the Q_DISABLE_COPY macro to the classes.

The CopyListContents code also needs to be simplified to not convert
from pointer to object to reference, because this causes an implicit
copy of the object. That means that all the object copy functions need
to be updated to use pointers instead of references.

General: Instances of QObject (and its subclasses) should be
considered unique items that cannot be copied.  Knowing that each
object has only one parent and an arbitrary number of children makes
it easier to understand this. Even if you "clone" the object by
setting all the member variables to the same values, you can't set the
parent and child variables to the same values without screwing up the
one-to-many relationship of parent to children.

See:
http://doc.qt.io/qt-5/qobject.html#Q_DISABLE_COPY
https://www.ics.com/designpatterns/book/qobject.html