Services API: Add/UpdateRecordSchedule, prevent DB Errors when timestamps aren't passed
Because they aren't required for SearchTypes other than None and Manual,
the start/end date/time, weren't being validated. If StartTime and EndTime
prarmeters weren't passed, then when an INSERT/UPDATE was done, the NULL
fields caused a DB Error and the API returned -1 ({"uint": "4294967295"}.)

This fix allows NULLs in the start/end date/time in the record table.

Fixes #13056