Save the alttext for a button in a themed menu.
This eliminates the compiler warning message, but does nothing to deal
with the lack of alttext support anywhere in the menu system.  The
alternative would be to remove support for parsing the alltext line in
themes.  The default theme in the git sources only has three entries
that specify alternate text, but it is unknown if it is used in other
themes.