Convert the function argument to a string, not the saved status.
This change is essentially a nop as one of the two callers to this
function already passes in the member status variable, and the other
didn't use the resulting value.  With this change, it is now possible
to convert a status enum to a string without having to first modify
the object status.  Add a new debug message to UpdateStatus that logs
the old and new status.