Comment out unused argument in CardUtil::CreateCardInput.
This function was created in 2011 (commit 32f11d9f) and has never used
the 'hostname' variable.  Given that this hasn't cause a problem in
the last six year, it would seem safe to remove, except that removing
it would require changes to the service contracts.  Its simpler for
now to just comment out the variable to eliminate the compiler
warning.