Comment out unused argument in PlayGroupConfig::PlayGroupConfig.
The 'label' variable was added to this function in May 2017 (commit
da287624) when the new playback group editor code was moved to
libmythtv.  It was used in the old version of the code from the
mythfrontend directory but isn't used in the libmythtv version.
Comment out the variable to prevent a compiler warning.