Services API: Log more Guide/GetChannelIcon errors
The primary fix is for folks that create a link between
~mythtv/.mythtv/channels and ~/.mythtv/channels, but
forget to make their channels directory writable by
the user running the backend. The link solution (rather
than running mythtv-setup as user mythtv) appears in
the Wiki and the Forum. When the new copy of the image
is attempted, e.g using Height=57, it failed.

Log and tell the user if the ChanId doesn't exist or is
marked invisible.

If the image can't be scaled, log the error. Happens now
when lyngsat returns text files rather than images.

Use upnp filtering to see the new logs.