Use all arguments passed to ChannelImporter::ShowManualChannelPopup.
This function has two unused arguments, 'parent' and 'title', that are
causing compiler warnings.  All callers to this function provide the
result of GetMythMainWindow() as the parent argument, so simply use
that variable instead of calling GetMythMainWindow() again inside of
the function.  The title argument can be used by switching to a
different MythDialogBox constructor that supports a title argument.