Fix doxygen complaints about macros before the keyword 'public'.
Doxygen gets upset and won't parse the file if the final macro
preceding the 'public' keyword isn't terminated with a semicolon.
Clearly this character isn't required by the compiler, but adding it
fixed doxygen's parsing of the files.  Oddly enough this also fixed
Emacs syntax highlighting code for these files as well.