ServicesAPI: add a couple of methods for Music
Add /Music/GetTrackList this method gets a list of all the music tracks
in the database. (VERY slow if you have a lot of tracks in the DB and
don't use Count to limit the list size).

Add /Music/GetTrack this gets the details of a single track.

This is just a proof of concept at the moment there is no sorting or
filtering and some things like coverart are missing.