Refs #13046 - Fix missing support for allow_single_step in spin boxes.
Many setup pages use the allow_single_step parameter for spin boxes
but that option was not implemented. This changes the use of
that parameter so that it will cause single step to be used for
the spin box left/right keys and the larger default step requested
to be used for the page up/down operation.