Prevent doxygen from parsing proposed future parameters.
The comment before the logServerStart function proposes a bunch of
possible new parameters, using a syntax that doxygen parser.  Doxygen
then complains because there is a whole set of documented parameters
that don't exist.  Escaping the first character on each of these
parameters prevents doxygen from trying to parse the line.