Escape special characters to prevent doxygen parse warnings.
There are several characters that doxygen attempts to parse as
code/text references and complains when it can't find a matching
target class/function/reference.  Most of these are resolved by
escaping the particular character so that it loses its special
meaning. A couple are resolved by adding a space so that doxygen
correctly parses the line of text.