Convert long list of ignored values into a default statement.
Commit c13b8361 in 2015 added two additional values to the MarkTypes
enum but didn't update the case statement in this file, and clang is
printing an 'enumeration not handled in switch' message to warn about
this omission.  Given that the function only cares about CUT_START and
CUT_END, it makes more sense to collapse the existing 24 (now 26)
ignored enumeration values into a single 'default' case.