Quiet clang 'enumeration not handled in switch' messages.
All of these functions for converting various items into strings
correctly handle the case where an item doesn't explicitly appear
within the switch statement.  This is done by code at the end of the
function that creates a generic response.  Add a default case to the
switch statement (that just breaks) to quiet the compilation warning.