Fix conditionally used member variables in the VideoOutputOMX class.
There are three member variable that are unconditionally defined and
initialized, but then only used if the string OSD_EGL is defined.
This causes clang to emit a warning error about private fields not
being used.  Move the declaration and initialization of these three
variables behind #ifdef statements to eliminate the compiler warnings
from clang.