Move an unused InfoMap printing function to a different file.
The clang compiler complains about an unused inline function for
printing out the contents of an InfoMap structure.  This seems like a
useful function to have in the code that could be reused elsewhere.
Move it from its current location to the file that defines the InfoMap
structure and take off the inline keyword (which eliminates the
compiler complaint).