Fix 'implicit conversion changes value' warning message.
This is an old problem of using signed character strings to hold
binary network data.  The network channel is 8-bit, but because the
string isn't declared as unsigned the compiler complains about the use
of the most significant bit.  Fix this warning by declaring the buffer
as holding unsigned characters, and then cast the call to the
writeDatagram function.  In an ideal world, the writeDatagram function
would take an uchar* argument, but at some point there will have to be
a cast to a char* as that's what the underlying QUdpSocket function
expects.