Fix 'max of value and unsigned zero' warning message.
The iterator in all of these functions is iterating over a structure
that contains two unsigned values, and is returning one of them.  It
is a truism that taking the max of an unsigned value and zero will
always return the unsigned value, therefore the call to max() is
superfluous.