Fix 'comparison of integers of different signs' warning message.
The 'rep' field in the lirc_state structure is unsigned, so convert
the parsing of that value to also use an unsigned variable.  In
lirc_read_string, use a size_t for the head and tail variables since
they're offsets into a buffer and are being compared to the result of
strlen(). In lirc_send_command, 'n' and 'data_n' should be the same
type.