Fix 'address of array will always evaluate to true' warning.
This appears to be a test to check whether both 1) a buffer has been
allocated and 2) the buffer contains data.  However, in this
particular case the data structure embeds the actual buffer (instead
of a pointer to a buffer) so that the first test is unnecessary. If
you have a structure, you have the embedded buffer.  Remove the test
to eliminate the compiler warning.