Move the themecache directory to a consistent location.
The cache for downloaded themes currently gets created in
<confdir>/cache/themecache/<themename>/themechooser, where themename
is the theme that you were using when to went into the theme chooser.
This means that there are likely multiple copies of this data on disk.
Move the cache to <confdir>/cache/themechooser so there is only ever
one copy present.