Make the Canadian weather scripts work again.
Fix the parsing code that is looking for the letter 'C' in the
temperature and dew-point lines.  This was unnecessarily escaped as
'\C' which causes Perl to complain about a no longer supported regex.
Removing the escaping solves the problem.

(cherry picked from commit 8f27146d1e3155bef2298b0033f804f23b8178bf)