Remove QueryInputDisplayName/kUnknownInputName nonsense.
As I remember it, when inputname was added to ProgramInfo, we didn't
want to further bloat the Myth protocol by passing it around too.
Instead, we made provisions to load it on demand from the database.
Later, someone else realized lots of database queries were happening
to get the inputname and slowing things down, so they added inputname
to the protocol.  This change [hopefully] makes sure inputname
initially gets set everywhere it should and removes all vestiges of
the load on demand support.