Change the ProgramInfo::ToMap() "input" item to include a shortened inputname.
The is to complement the existing "card" item which provides a short,
combined version of input number for recording/will record entries and
recording status for won't record entries.  The difference now is that
the "input" variation will show the last two characters of the
configured input display name instead of the input number.  This can
be more useful in configurations with large numbers of inputs and high
max recordings, multirec settings.