ProgramDetails: Provide "Custom Priority" adjustment details as second level info.
The Custom (power) priorities are a powerful tool, but it can be difficult
to diagnose why they don't seem to have the desired affect.  Detailing the
priorities rules which match, and their effect on the recording priority
makes it much easier for the user to see why scheduling decisions are being
made.