ttvdb: Various fixes
1. ttvdb: Update "Fixer Upper" test cases with wildcards
"Lastupdated" field has changed, and seems likely to change, so make it a
wildcard match. "Thumbauthor" field has also changed.
2. Update "NOVA: Eclipse Over America" test case
This appears to have moved from episode 18 to episode 11 and had updates
to the description and other metadata.
3. Ignore banner of just  "http://thetvdb.com/banners/"
4. ignore user rating of 0
5. correct handingly of genre
They are a list of "<item>..</item>".
6. Add new test case: ttvdb.py -l en -a GB -C 330432
This had some regressions compared to v29.0 which are now fixed, so add a test case.

Fixes #13130

Signed-off-by: Peter Bennett <pbennett@mythtv.org>