Use the define property flags when creating a RecordingInfo object.
This code used a hard coded shift for adding the subtitle flags to the
properties field, and that value hasn't kept up with the addition of
new flags.  Change the hard coded number to a constant defined in the
same location as the flags.  One reported visible side effect of this
bug is that mythweb would indicate that most recording are damaged.
This is the result of the subtitle "hard of hearing" flag being
wrongly shifted into the position of the video damaged flag.