Parse subclass specific tags before calling base class parser.
Modify the code to parse the subclass tags first, and only call the
base class parser on unrecognized tags.  This prevents the base class
parser from complaining about all subclass tags.  Moving the call to
the base function exposes the problem that both the base class and the
subclass are coded to match the "font" tag.  Remove the subclass match
as the parser in the base class is the correct parser for that tag
(and the subclass match has never happened since the base class always
returns true for the "font" tag preventing the subclass match.)