Redesigned Previously Recorded Screen
Implements a new Previously Recorded screen. The left column is
switchable between month dates and titles, and the right column
gives a list of previously recorded episodes of the show selected
on the left, or a list of recordings for the month selected on the
left.

The old Previously Recorded screen is still available via the
Default menu.

Fixes #13135