Fix cppcheck "Condition 'vtp==0' is always true" warning.
The code that would set this variable to a non-zero value is commented
out, so this is a valid warning.  Paring the function down to the only
possible code path would yield a function that always returns NULL.
Instead, #ifdef out the code in case someone ever gets around to
working on it.