Fix warning for conditionally used variables.
There are a couple of places in this file where function arguments are
only used if the file is compiled with USING_FIREWIRE defined.  When
its not defined the compiler emits a warning error for an unused
variable.  Add some code to quiet the compiler only when not
supporting firewire.