Fix compilation error in Fedora Rawhide.
A recent change to the kernel headers (for 4.15.0) changed the most of
the fields in the iso_primary_descriptor data structure from signed
character arrays to unsigned character arrays.  The one used by MythTV
needs to be cast back to signed characters so that it can continue to
be used to initialize a QString.

(cherry picked from commit 90f9457b74f1ccd0df31c825b1e721203e5075ff)