Fix linking of test binaries on Ubuntu 17.10.
The default linker behavior has changed in Ubuntu 17.10.  The old
behavior was to emit a DT_RPATH tag (which has a similar effect to the
LD_LIBRARY_PATH environment variable) which applied both to the search
path of the binaries and the search path of dependent libraries.  The
new behavior is to emit a DT_RUNPATH line which only applies to the
search path of the binaries and not to dependent libraries.  Adding
-Wl,--disable-new-dtags will restore the old behavior.  The new
behavior is considered "more correct" so this will probably have to be
revisited in the future.