Add missing else cases in several tvremoteutil functions.
In the "called from backend context" cases, a missing TVRec would
cause the backend to try to connect back to itself which isn't
allowed.

Note that the error case for RemoteIsBusy is handled differently
between the backend and frontend cases.  In the backend case, we don't
want a missing input block other actions.