Get WSDL working again after QObject cleanups.
Don't strip the '*' from type names when looking up types in the type
system.  This worked before because the old code registered both
"ObjectName" and "ObjectName*" in the type system.  The cleaned up
code no longer registers "ObjectName" because it no longer has a
public copy constructor (which is a requirement for objects derived
from a QObject).