Get XSD working again after QObject cleanups.
When adding "DTC::" to the start of the type name before a lookup, now
also add '*' to the end of the name.  This worked before because the
old code registered both "ObjectName" and "ObjectName*" in the type
system.  The cleaned up code no longer registers "ObjectName" because
it no longer has a public copy constructor (which is a requirement for
objects derived from a QObject).