Remove usage of QString::null when used to test for a null string.
The QString::null value was officially deprecated in Qt 5.9.
According to the commit message where it was deprecated, this value
was a worrisome holdover from Qt3 and should be eliminated from code.
This patch removes all instances where the value was used to test
whether a string contains a value.  The commit messages says to use
QString::isNull, but the Qt documentation says that what you probably
really want is QString::isEmpty.  This patch uses the latter, except
in places where it seems obvious that the code writer truly thought
the string might be null.  The remaining instances that check for null
probably need to be revisited.