Don't assign newly declared QStrings the value QString::null.
The QString::null value was officially deprecated in Qt 5.9.
According to the commit message where it was deprecated, this value
was a worrisome holdover from Qt3 and should be eliminated from code.
Assigning newly declared strings the value of QString::null is
essentially redundant, and will stop working in the future, so remove
any such assignment.