Don't use QString::null in class constructors.
The QString::null value was officially deprecated in Qt 5.9.
According to the commit message where it was deprecated, this value
was a worrisome holdover from Qt3 and should be eliminated from code.
Constructing new instances of class strings using QString::null is
essentially the same as using the default constructor, so remove any
such initialization.