Replace assigning QString::null to a string with a call to clear().
The QString::null value was officially deprecated in Qt 5.9.
According to the commit message where it was deprecated, this value
was a worrisome holdover from Qt3 and should be eliminated from code.
Replacing assignment of QString::null with a call to clear() was also
used in the commit to Qt that deprecated QString::null.