Replace deprecated qBinaryFind function with std::lower_bound.
According to the Qt5.0 documentation, you should use standard library
functions instead of the Qt supplied qBinaryFind function.  The docs
recommend std::binary_search (for testing if an item is in a
container) or std::lower_bound (for finding an item in a container).
The old function has been deprecated.