Flag the RecordingInfo from ProgramInfo constructor as explicit.
This pointed out a number of places where the code was calling a
function and passing a ProgramInfo structure when the function
expected a RecordingInfo structure expected.  Instead of changing the
callers, it seemed simpler to overload the functions being called and
convert in the called functions.  This constructor was pointed out by
cppcheck.