Fix cppcheck signedness complaint in HistogramAnalyzer.cpp.
Cppcheck complained that a variable being read with a '%2x' format
needed to be unsigned instead of signed.  This patch changes the array
used for reading to be unsigned instead of signed, and also removes
the test for less than zero (since that can't happen with an unsigned
variable.)  This code is being used to read an array of uchar values
(0-255) printed earlier, so there should be no problem with this
change.