Fix cppcheck "constructor with 1 argument that isn't explicit" warnings.
Fix this set of constructor warnings for the classes in the file
dtvconfparserhelpers.h.  Mark the constructors as explicit, and
tighten up the parameters by requiring that the argument be a class
specific enum instead of a generic integer.  These enums are all
copies of enums from a system header files, so explicitly link them to
the system header enums.  This patch also adds 'operator=' functions
to the classes for both the class enum and the system header enum,
since both are used interchangeably in the code.  Add a couple of
explicit conversions to DTVTunerType where needed.