Qt5.8 QDateTime: Update mythnews to use the new API.
Replace the call to QDateTime::toTime_t with a call to
QDateTime::toSecsSinceEpoch.  Make a little code tweak where this
function could be called three times in a row on the same value.
Change a direct comparison to a zero timestamp to use the isValid()
function instead.

The QDateTime {to,from}Time_t functions were obsoleted in Qt5.8 and
replaced by {to,from}SecsSinceEpoch functions.  The two sets of
functions perform the same operation, only the latter set takes a
wider range of values than the former (uint vs. qint64).