Qt5.8 QDateTime: Update the ImageItem class to use the new API.
Change two of the fields in this class from uint -> qint64.  Replace
the calls to QDateTime::{to,from}Time_t with the corresponding calls
from the new API.  Replace "currentMSecsSinceEpoch() / 1000" with
"currentSecsSinceEpoch()".

The QDateTime {to,from}Time_t functions were obsoleted in Qt5.8 and
replaced by {to,from}SecsSinceEpoch functions.  The two sets of
functions perform the same operation, only the latter set takes a
wider range of values than the former (uint vs. qint64).