Qt5.8 QDateTime: Update {atsc,scte}tables.h et al to use the new API.
This is a straight substitution of eight calls to ::toTime_t with
calls to ::toSecsSinceEpoch.

The QDateTime {to,from}Time_t functions were obsoleted in Qt5.8 and
replaced by {to,from}SecsSinceEpoch functions.  The two sets of
functions perform the same operation, only the latter set takes a
wider range of values than the former (uint vs. qint64).