Qt5.8 QDateTime: Update UPnpCDSTv::LoadRecordings to use the new API.
This is a substitution of two calls to ::toTime_t and a millisecond
conversion, with just two calls to ::toMSecsSinceEpoch.

The QDateTime {to,from}Time_t functions were obsoleted in Qt5.8 and
replaced by {to,from}SecsSinceEpoch functions.  The two sets of
functions perform the same operation, only the latter set takes a
wider range of values than the former (uint vs. qint64)