Qt5.8 QDateTime: Update image caching to use the new API.
This is a straight substitution of calls to ::toTime_t with calls to
::toSecsSinceEpoch, and changing variables from uint to qint64.  There
is also one change to replace ::toMSecsSinceEpoch()/1000 with
toSecsSinceEpoch().

The QDateTime {to,from}Time_t functions were obsoleted in Qt5.8 and
replaced by {to,from}SecsSinceEpoch functions.  The two sets of
functions perform the same operation, only the latter set takes a
wider range of values than the former (uint vs. qint64).