Qt5.8 QDateTime: Update EIT scanner to use the new API.
This is a straight substitution of calls to ::toTime_t with calls to
::toSecsSinceEpoch.  All calculations are still done with 32bit
values, as this assumption is built into the EIT code.

The QDateTime {to,from}Time_t functions were obsoleted in Qt5.8 and
replaced by {to,from}SecsSinceEpoch functions.  The two sets of
functions perform the same operation, only the latter set takes a
wider range of values than the former (uint vs. qint64).