Qt5.8 QDateTime: Update LiveTVChain to use the new API.
This is a straight substitution of a call to ::setTime_t with a call
to ::setSecsSinceEpoch.

The QDateTime {to,from}Time_t functions were obsoleted in Qt5.8 and
replaced by {to,from}SecsSinceEpoch functions.  The two sets of
functions perform the same operation, only the latter set takes a
wider range of values than the former (uint vs. qint64).