Qt5.8 QDateTime: Update protocol command handling for two PIXMAP commands.
Update the handling for the QUERY_PIXMAP_LASTMODIFIED and
QUERY_PIXMAP_GET_IF_MODIFIED protocol commands.  This is mostly a
straight substitution of calls to ::{to,from}Time_t with calls to
::{to,from}SecsSinceEpoch.  It has a couple of additional validity
checks, and continues to use the values "BAD" and ((uint)-1) for
invalid QDateTime values.

The QDateTime {to,from}Time_t functions were obsoleted in Qt5.8 and
replaced by {to,from}SecsSinceEpoch functions.  The two sets of
functions perform the same operation, only the latter set takes a
wider range of values than the former (uint vs. qint64).