Qt5.8 QDateTime: Update protocol handling for QUERY_SG_FILEQUERY command.
Update the handling for the QUERY_SG_FILEQUERY protocol command.  This
is mostly a straight substitution of calls to ::{to,from}Time_t with
calls to ::{to,from}SecsSinceEpoch.  It has a couple of additional
validity checks, and continues to use the value ((uint)-1) for invalid
QDateTime values.

The QDateTime {to,from}Time_t functions were obsoleted in Qt5.8 and
replaced by {to,from}SecsSinceEpoch functions.  The two sets of
functions perform the same operation, only the latter set takes a
wider range of values than the former (uint vs. qint64).